Split into three files and got rid of extra sandbox files.
